Refactor default ApiRequestExecutor implementation
- Move to `DefaultApiRequestExecutor`
- Inject `CoroutineContext` into `DefaultApiRequestExecutor`

This makes `DefaultApiRequestExecutor` responsible for switching
to the correct context.
